Add useFetch
